refactor(training): rename Success modal to TrainingSuccess (PR review)
Names the modal for its training-module context throughout — view-model,
screen controller, Alloy presenter, view/tss, registry and specs. Also
extracts the copy-pasted makeWidget/makeBackboneTarget fakes into a shared
test/fixtures/fakeWidgets.js and uses it here; migrating the other specs is
carded (WB-227).
