+++ walta-app/app/controllers/TrainingSuccess.js
@@ -1,4 +1,4 @@
 // Presenter shell only. All behaviour lives in the Titanium-free
-// lib/mvvm/controllers/Success screen controller, which View.openModal
+// lib/mvvm/controllers/TrainingSuccess screen controller, which View.openModal
 // instantiates with this controller's widgets ($) and a close callback.
 // Nothing to wire here.
