+++ walta-app/app/styles/TrainingSuccess.tss
